Add landing page and move solver
